requirements.txt
pandas
openai 
python-dotenv
flask 
matplotlib
numpy
pillow
